+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this dbapi driver will be documented in this file.
   * New optional keyword parameter `properties: PropertiesDict | None` on `Cursor.execute()` and related methods to allow callers to provide [statement execution properties](https://docs.confluent.io/cloud/current/flink/reference/statements/set.html#table-options). Note: connection or cursor-level properties for default catalog, database, and execution mode cannot be overridden by this parameter.
 
 ### Changed
-  * Respelled the `connect()` parameter `dbname` to `database`.
+  * Respelled the `connect()` parameter `dbname` to `database`. The old spelling `dbname` is deprecated and will be removed in after one release cycle.
 
 ## 0.1.x